requirements.txt
mag4
plotly.express
pandas